Extract chip states as separate component
